factored out isScanningAllowed() method
